requirements.txt
streamlit
openai
snowflake-connector-python
python-dotenv
pandas
